[FIX] 추가,  삭제 버그 수정 및 컴포넌트 분리
