+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_SERVER_URL = http://3.38.94.39:8080/api
+REACT_APP_SERVER_URL = http://43.200.121.108:8080/api
